+++ wcqtlib/data/dataset.py
@@ -8,7 +8,9 @@ import logging
 import os
 import pandas
 import requests
+from sklearn.cross_validation import train_test_split
 import sys
+
 
 import wcqtlib.common.config as C
 
@@ -119,11 +121,11 @@ class Dataset(object):
             logger.warning("No observations to validate.")
             return False
 
-    def view(self, dataset):
+    def view(self, dataset_filter):
         """Returns a copy of the analyzer pointing to the desired dataset.
         Parameters
         ----------
-        dataset : str
+        dataset_filter : str
             String in ["rwc", "uiowa", "philharmonia"] which is
             the items in the dataset to return.
 
@@ -131,17 +133,68 @@ class Dataset(object):
         -------
         """
         thecopy = copy.copy(self.to_df())
-        ds_view = thecopy[thecopy["dataset"] == dataset]
+        ds_view = thecopy[thecopy["dataset"] == dataset_filter]
         return ds_view
+
+    def get_train_val_split(self, test_set, train_val_split=0.2,
+                            max_files_per_class=None):
+        """Returns Datasets for train and validation constructed
+        from the datasets not in the test_set, and split with
+        the ratio train_val_split.
+
+         * First selects from only the datasets given in datasets.
+         * Then **for each instrument** (so the distribution from
+             each instrument doesn't change)
+            * train_test_split to generate training and validation sets.
+            * if max_files_per_class, also then restrict the training set to
+                a maximum of that number of files for each train and test
+
+        Parameters
+        ----------
+        test_set : str
+            String in ["rwc", "uiowa", "philharmonia"] which selects
+            the hold-out-set to be used for testing.
+
+        Returns
+        -------
+        train_df, valid_df : pandas.DataFrame
+            DataFrames referencing the files for train and validation.
+        """
+        df = self.to_df()
+        datasets = set(df["dataset"].unique()) - set([test_set])
+        search_df = df[df["dataset"].isin(datasets)]
+
+        selected_instruments_train = []
+        selected_instruments_valid = []
+        for instrument in search_df["instrument"].unique():
+            instrument_df = search_df[search_df["instrument"] == instrument]
+
+            if len(instrument_df) < 2:
+                logger.warning("Instrument {} doesn't haven enough samples "
+                               "to split.".format(instrument))
+                continue
+
+            traindf, validdf = train_test_split(
+                instrument_df, test_size=train_val_split)
+
+            if max_files_per_class:
+                traindf = traindf.sample(n=max_files_per_class)
+
+            selected_instruments_train.append(traindf)
+            selected_instruments_valid.append(validdf)
+
+        return pandas.concat(selected_instruments_train), \
+            pandas.concat(selected_instruments_valid)
 
 
 class TinyDataset(Dataset):
-    ROOT_PATH = os.path.abspath(os.path.join(os.path.dirname(__file__),
+    ROOT_PATH = os.path.abspath(
+        os.path.join(os.path.dirname(__file__),
         os.pardir, os.pardir, "tests", "tinydata"))
-    DS_FILE = os.path.join(self.ROOT_PATH, "tinydata.json")
+    DS_FILE = os.path.join(ROOT_PATH, "tinydata.json")
 
     @classmethod
-    def read_json(cls, json_path=DS_FILE):
+    def load(cls, json_path=DS_FILE):
         with open(json_path, 'r') as fh:
             data = json.load(fh)
             # Update the paths to full paths.
